added view movie endpoint, refactored and organized code a little bit

+++ src/main/java/com/grono/moviematchbackend/repository/MovieRepository.java
@@ -1,14 +1,13 @@
 package com.grono.moviematchbackend.repository;
 
-import com.grono.moviematchbackend.model.Movie;
+import com.grono.moviematchbackend.model.movie.Movie;
 import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.mongodb.repository.Query;
 
 import java.util.Optional;
 
 public interface MovieRepository extends MongoRepository<Movie, String> {
 
 
-    Optional<Movie> findMovieByMovieId(Long movieId);
+    Optional<Movie> findMovieByMovieId(Integer movieId);
 
 }
